Connected C++ to Python. Hello world is working!!!
