+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = ["iAirLab <iairlab@yandex.ru>"]
 license = "BSD-3-Clause"
 # readme = "README.rst"
 # readme = "README.md"
-build = "build.py"
+#build = "build.py"
 #generate-setup-file = false
 
 [tool.poetry.dependencies]
@@ -25,5 +25,5 @@ pathos = ">=0.3.0,<0.3.1"
 
 
 [build-system]
-requires = ["poetry>=0.12", "wheel", "setuptools-cpp"]
+requires = ["poetry>=0.12", "wheel", "setuptools"]
 build-backend = "poetry.masonry.api"
